+++ crates/aviutl2/src/input/binding.rs
@@ -141,8 +141,6 @@ pub enum InputType {
     Audio,
     /// 動画と音声の両方。
     Both,
-    /// 画像のみ。
-    Image,
 }
 
 impl InputType {
@@ -151,7 +149,6 @@ impl InputType {
             InputType::Video => 1,
             InputType::Audio => 2,
             InputType::Both => 3,
-            InputType::Image => 4,
         }
     }
 }